Number of improvements to the main loop, improving performance:
- job.exception() blocks until the job is done. Because of this, the
  main loop would always wait until all jobs were finished before
  executing the next iteration

- Because of this, the job > 180s code path wasn't reachable
  This path still had some bugs that were fixed

- Added a queue in front of the running jobs. This way, we can already
  retrieve the next job while the previous one is still running, hiding
  the job-pop latency

- log timestamps with microsecond precision + added debug logging for
  performance tuning

GPU's are actually not great at running multiple workloads at once; On a
3090 I see ~20% total throughput drop as soon as the second job starts.
With these optimisations, it should be possible to run the worker with
max_threads = 1 for optimal performance.

Even for higher thread counts, the job.exception block prevented these
threads from getting the highest utilisation of the GPU,
so even in that case performance should be significantly better.