Add optional A1111 style prompt weight parsing
As briefly discussed on Discord:
https://discord.com/channels/781145214752129095/1027697691628679189/1070983826022613063

A lot of people seem to be still using A1111 style prompt weights.
Let's not mention it in the documentation (we like our own style:1.2),
but it would still be good to accomodate these users / increase
reproducibility compared to A1111 / etc.

This commit adds a A1111 to Nataili style weights parser.